SOBI Updates
Updates for the SOBI_Wrapper function: the whitening is now done in Julia, including regularization of the matrix to take the eigendecomposition of. The jointDiag package in R is now used. Previously, the JADE package's SOBI function was used; the lack of regularization of the square-root of the inverse covariance was causing issues.